Add files via upload
About the Dataset

The timeline being compared in this study is between the Years 2008 to the Year 2022. This data has been collected from The World Bank Databases on the topic of Environment, Social and Governance (ESG) Data-data#). The dataset represented all 193 recognized countries in total but this study is mainly focused on five countries for research:

United States: It is the third-largest country in the world by land mass area with around 9.8 million square kilometers.

India: It is the seventh-largest country in the world by land mass area with about 3.29 million square kilometers.

Indonesia: It is the largest archipelagic country and the fourth most populous country in the world. However, its total land area is approximately 1.9 million square kilometers.

Australia: It is pertinent to study the environmental impact on the economy due to its significant dependence on natural resources, especially in mining and agriculture. The country faces challenges in managing economic growth while addressing environmental concerns related to land use, biodiversity, and climate change.

Netherlands: This country provides valuable insights into managing the environmental impact on the economy through sustainable practices. As a highly developed nation and with a focus on innovation & environmental policies, the Netherlands demonstrates how advanced economies can balance economic prosperity with responsible resource management and conservation efforts.

Who can benefit from this research?

- This analysis is beneficial for investors of upcoming industries and builders of new factories, stakeholders, and consumers who will be using the resources. Such data can also help people acknowledge their usage of natural resources and take proper precautions to reduce the overuse of these resources which leads to the degradation of nature and the environment.

- Analyzing and reading such data is also helpful for governments to take informed actions in favor of the environment and limit the usage of chemicals and toxic behaviors of these chemicals